requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
PyPDF2
python-dotenv

